Add djoser and modify permissions

+++ restaurant/models.py
@@ -14,7 +14,7 @@ class Booking(models.Model):
 class MenuItem(models.Model):
     title = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=5, decimal_places=2)
-    inventory = models.IntegerField()
+    inventory = models.SmallIntegerField()
 
     def __str__(self):
         return self.title
